5.4.19-20241103_17309482

+++ polardbx-test/src/test/java/com/alibaba/polardbx/qatest/ddl/auto/ddl/ForeignKeyDdlTest.java
@@ -599,6 +599,10 @@ public class ForeignKeyDdlTest extends PartitionTestBase {
 
     @Test
     public void CreateTableDifferentCharsetWithFkReferred() throws SQLException {
+        if (isMySQL80()) {
+            return;
+        }
+
         JdbcUtil.executeUpdateSuccess(tddlConnection, "SET ENABLE_FOREIGN_KEY = true");
 
         dropTableIfExists("charset_c");
